Reorganize string extensions
